pyproject.toml
[project]
name = "work-logger"
version = "0.1.0"
description = "Work logger with pywebview gui."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pywebview>=6.1",
]
